add index.js in components, refactor code
