+++ src/components/index.js
@@ -1,0 +1,5 @@
+export { default as Hero } from "./Hero.svelte";
+export { default as ScoreCard } from "./ScoreCard.svelte";
+export { default as AddPlayer } from "./AddPlayer.svelte";
+export { default as Button } from "./Button.svelte";
+export { default as Footer } from "./Footer.svelte";
